Adding endpoints for Audience Rule creation
Summary:
# Context:
Injecting a new rule creation endpoint will lead to creation of an empty subscriber list in Ads Manager.

# Details:
This code is a Node.js Lambda function that handles HTTP requests for an audience rule management system. The function uses the AWS SDK to interact with AWS services such as Secrets Manager and RDS. It also uses the mysql and axios libraries to interact with a MySQL database and make HTTP requests to a third-party API, respectively.

- If the HTTP method is GET, the function queries the database for all audience rules using the queryDatabase function and returns the results as a JSON response.
- If the HTTP method is POST, the function parses the request body to extract the name, include, and exclude parameters. It then generates a SQL query for the audience rule using the createSqlQueryForAudienceRule function and inserts the rule into the database using the queryDatabase function. The function then creates a subscriber list using the createSubscriberList function and updates the database with the subscriber list ID. Finally, the function selects the newly inserted row from the database and returns it as a JSON response.
- If the HTTP method is DELETE, the function deletes all rows from the audience_rules table in the database using the queryDatabase function and returns a success message as a JSON response. If the HTTP method is not GET, POST, or DELETE, the function returns a "Method Not Allowed" error message.

In the finally block, the function closes the database connection if it was successfully established.

Overall, this code provides a way to manage audience rules for a marketing campaign by allowing users to create, read, update, and delete rules.

Reviewed By: vitorfalcaor

Differential Revision: D51515978

fbshipit-source-id: 74388129bdb227451ac5d60bbe6f3348313b6c2d