Add POST and GET endpoint for processing meta webhook
Summary:
This diff adds a new ednpoint to receive meta webhooks on AWS. This endpoint has GET and POST methods implemented.

It also adds two new fields to the `iac.yaml` file: 'WebhookEndpointName' and 'WebhookStageName'. These fields are designed to define the endpoint and stage for processing webhook notifications.

The 'WebhookEndpointName' is used to specify the name of the endpoint, such as 'subscribe-notifications', and the 'WebhookStageName' specifies the stage of the endpoint, like 'v1'.

These configurations are essential for setting up the webhook endpoint in the AWS API Gateway.

{F1772057177}

Reviewed By: matheusmso

Differential Revision: D60236143

fbshipit-source-id: 8087932f9878fda61a6fa66c372423e906ae6041