Add new lambda function to handle webhooks
Summary:
This diff adds a new lambda function to handle webhooks for the WMG (Webhook Management Gateway) project. The lambda function is named `WebhookProcessing`.

The function is a dummy function that does not process the message and return 200. The function is configured to run in the same VPC created by WMG.

It also has a dead letter queue configured to handle any failures.

{F1772056722}

Reviewed By: matheusmso

Differential Revision: D60236142

fbshipit-source-id: 7919ce01ac4576a5e3300e0ec5eed6dfb11f6d93