Adding database backups retention period parameter
Summary:
With the  use of WMG at a large scale, the database tables get to store a high volume of important data regarding messages measurement. At this point, it is important to retain DB snapshots (backups).

This diff includes a new parameter to the IAC file: the number of days to retain snapshots automatically generated in RDS. The default value is 7 (days), while the maximum is 35 and the minimum is 0 (backups disabled).

Reviewed By: matheusmso

Differential Revision: D65564350

fbshipit-source-id: 02aa6b6ab11d179a64ce7d91ccee3311dc9b4d19