Consider specific templates to apply lift study flow
Summary:
This diff introduces a new criteria to apply the lift study logic in the Router.

Before this diff: whenever a lift study is active, any template messages will trigger the lift study flow for assigning phone numbers to groups and discarding messages to users assigned to control group.

Issue: consider one using WMG to send messages in a solution that cannot separate proactive and reactive conversations, and has any templates other than the one that starts conversations. In this case, users in control group would have their organic experiences impacted by the missing templates that would be discarded.

After this diff: it will be possible to link template names to a lift study. Router starts considering the template names to decide whether to apply or not the lift study logic.

Reviewed By: mirellasilva

Differential Revision: D69727127

fbshipit-source-id: b2d03ac54091165fa2b83d5fa5d4582030e3a0a4