Add control over templates considered to the study
Summary:
This diff introduces a new dimension for lift study control.

Before this diff: whenever a lift study is active, any template messages will trigger the lift study flow for assigning phone numbers to groups and discarding messages to users assigned to control group.

Issue: consider one using WMG to send messages in a solution that cannot separate proactive and reactive conversations, and has any templates other than the one that starts conversations. In this case, users in control group would have their organic experiences impacted by the missing templates that would be discarded.

After this diff: it will be possible to link template names to a lift study. The template names are stored as a string with comma-separated values.

After the diff D69727127 the router will consider the template names to decide whether to apply or not the lift study logic.

Reviewed By: mirellasilva

Differential Revision: D69726863

fbshipit-source-id: 1daf9c50cad4281b3fe5ed531b0ff8602153f2c1